api.js moved to upper folder (no longer in screens folder)
